fix(db): migrate SERIAL to BIGSERIAL and update all ID types to int64

+++ server-go/models/models.go
@@ -4,7 +4,7 @@ import "time"
 
 // User represents the users table structure
 type User struct {
-	ID                      int        `json:"id"`
+	ID                      int64      `json:"id"`
 	Login                   string     `json:"login"`
 	Username                string     `json:"username"`
 	Password                string     `json:"-"` // Never expose password hash in JSON
@@ -19,38 +19,38 @@ type User struct {
 
 // PrivateChat represents the private_chats table structure
 type PrivateChat struct {
-	ID        int       `json:"id"`
-	User1ID   int       `json:"user1Id"`
-	User2ID   int       `json:"user2Id"`
+	ID        int64     `json:"id"`
+	User1ID   int64     `json:"user1Id"`
+	User2ID   int64     `json:"user2Id"`
 	CreatedAt time.Time `json:"createdAt"`
 }
 
 // Message represents the messages table structure
 type Message struct {
-	ID               int        `json:"id"`
-	UserID           int        `json:"userId"`
+	ID               int64      `json:"id"`
+	UserID           int64      `json:"userId"`
 	Username         string     `json:"username"`
 	Message          string     `json:"message"`
-	ChatID           *int       `json:"chatId"`
+	ChatID           *int64     `json:"chatId"`
 	ChatType         string     `json:"chatType"` // "group" or "private"
 	Nonce            *string    `json:"nonce"`
 	SenderPublicKey  *string    `json:"senderPublicKey"`
-	ReplyToMessageID *int       `json:"replyToMessageId"`
-	RecipientID      *int       `json:"recipientId"`
+	ReplyToMessageID *int64     `json:"replyToMessageId"`
+	RecipientID      *int64     `json:"recipientId"`
 	IsEncrypted      int        `json:"isEncrypted"` // 0 = plain, 1 = E2EE
 	Date             time.Time  `json:"date"`
 }
 
 // ReplyPreview represents the preview metadata of a referenced reply message
 type ReplyPreview struct {
-	ID              int     `json:"id"`
-	UserID          int     `json:"userId"`
+	ID              int64   `json:"id"`
+	UserID          int64   `json:"userId"`
 	Username        string  `json:"username"`
 	Message         string  `json:"message"`
 	IsEncrypted     int     `json:"isEncrypted"`
 	Nonce           *string `json:"nonce"`
 	SenderPublicKey *string `json:"senderPublicKey"`
-	RecipientID     *int    `json:"recipientId"`
+	RecipientID     *int64  `json:"recipientId"`
 	MediaType       *string `json:"mediaType"`
 	MediaName       *string `json:"mediaName"`
 }
